Add recommends option to packages
Recommends is needed for Yocto to be able to use apk. If package A
recommends package B and apk is adding package A, then package B
will also be added. However if there's an error selecting B (for
example if it's not available) apk will not throw an error but just
continue and exit with success. A recommends is a non critical
depends.