database: move lock file from /lib/apk to /run/apk
/lib is only supposed to hold essential shared libraries and kernel
modules. The lock file is part of the state information of the
application. And could also live under /var/lib/apk/db/lock. However,
given that there is a directory specific to hold runtime data, let's put
it in there.

In addition, we don't need to check for APK_OPENF_CREATE when
requesting the lock file, since the directory where the lock lives now
is not related to the DB

This is a partial revert of 57de8d0c5de43dd8c1d7b76a429b34a7e5ddd7b0,
but the problem identified there should not be an issue when writing
directly to /run