database: move the root of the DB from /lib/apk to /usr/lib/apk
This is to allow and easier support transition to the /usr merge in
alpine. In addition /lib is only supposed to hold essential shared
libraries to boot the system and kernel modules. It is a wonder
whether apk-tools is essential, but any system can certainly run
without it.

Relates https://gitlab.alpinelinux.org/alpine/tsc/-/issues/73

Relates #10778