lua: fix install path with meson
The install_dir for luaapk was previously not in the search path of lua
modules. Additionaly the filename (libluaapk.so) did not match the name of
the module which prevented the lua module from being loaded correctly.

Due to meson not accepting targets with the same name I had to split the
lua module into a subdirectory.