print notice when a file is written to .apk-new
If there is no indication of a .apk-new file being created, users may
mistakenly believe that the installation was successful, even if apk did not
write to the file. With this notice, the user will be informed that further
action is required (such as using update-conf).

Inspiration by pacman which prints the following message:

warning: /etc/os-release installed as /etc/os-release.pacnew